+++ CHANGELOG.md
@@ -24,6 +24,7 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 ### Added
 
 - More dash icon size (16px, 22px and 24px).
+- Overview spacing size.
 - Screen recording indicator visibility.
 - Screen sharing indicator visibility.
 - Switcher popup delay by [@koplo199](https://gitlab.gnome.org/koplo199).